🐛 fix(vault-events): rename to events.svelte.ts and update codex-review
