+++ apps/web/src/lib/stores/vault/events.svelte.ts
